+++ lib/presentation/first_page.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
+import 'second_page.dart';
+
 class FirstPage extends StatelessWidget {
   static final route = "/first";
 
@@ -55,22 +57,4 @@ class FirstPage extends StatelessWidget {
               )
             ],
           ));
-}
-
-class SecondPage extends StatelessWidget {
-  static final route = "/second";
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        body: Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          ElevatedButton(child: Text("Return 42"), onPressed: () => Navigator.pop(context, "42")),
-          ElevatedButton(child: Text("Return AbErVaLlG"), onPressed: () => Navigator.pop(context, "AbErVaLlG")),
-        ],
-      ),
-    ));
-  }
 }
